+++ build.rs
@@ -8,7 +8,7 @@ fn main() {
     println!("cargo:rerun-if-changed=proto/google/pubsub/v1/pubsub.proto");
 
     let mut prost_build = prost_build::Config::new();
-    prost_build.btree_map(&["."]);
+    prost_build.btree_map(["."]);
 
     tonic_build::configure()
         .compile_with_config(