+++ src/lib.rs
@@ -33,14 +33,16 @@
 //!
 //! Users can choose appropriate http clients to align with their runtime.
 //!
-//! Based on the feature enabled. The only client available is surf, feel free to implement
+//! Based on the feature enabled. The only client available is reqwest, feel free to implement
 //! other http clients.
 //!
 //! Note that async http clients may need specific runtime otherwise it will panic. User should make
 //! sure the http client is running in appropriate runime.
 //!
 //! Users can always use their own http clients by implementing `HttpClient` trait.
-//!
+
+#![deny(unused_crate_dependencies)]
+
 pub(crate) mod dd_proto {
     include!(concat!(env!("OUT_DIR"), "/dd_trace.rs"));
 }